Moved PVResultsDataFrame to deprecated position.
Essentially this commit means that all functionality from
PVResultsDataFrame has been moved to the parent class,
PVPerformanceResults, to match PVFunctionalResults class.

The next thing to do is figure out how to remove all update functions
from both PVPerformanceResults and PVFunctionalResults and place them
in PVResultsHandler in a process-safe way. The current plan is to use
a shelve.Shelf and then when it's time to calculate use a calculator
class (e.g. PVResultsDataFrameCalculator).